Extracted dynamic command support from Composer to BaseSimpleCommand
The way we can now run all composer commands with per-command options
is useful in other places.  Therefor, this functionality is now
extracted into BaseSimpleCommand which all such commands can extends.
